Removed raw c pointers and replaced them with std::vector.
In order to avoid the disasters with C pointers now the
C++ interface uses std::vector which also carries the length
of the vector so that sanity checks can be done. The python
interface hasn't changed but under the hood it uses now
also std::vector.

With the length checks in place now an execption is thrown
if the preditctor or error arrays have a different length
as specificed in the constructor.